Created separate files for different tests
Relates #16

Relates #18
